+++ src/dataframe/insertion.jl
@@ -333,9 +333,10 @@ it must be a `Tuple` or an `AbstractArray`
 and columns are matched by order of appearance. In this case `row` must contain
 the same number of elements as the number of columns in `df`.
 
-If `row` is a `DataFrameRow`, `NamedTuple` or `AbstractDict` then
-values in `row` are matched to columns in `df` based on names. The exact behavior
-depends on the `cols` argument value in the following way:
+If `row` is a `DataFrameRow`, `NamedTuple`, `AbstractDict`, or
+`Tables.AbstractRow` then values in `row` are matched to columns in `df` based
+on names. The exact behavior depends on the `cols` argument value in the
+following way:
 * If `cols == :setequal` (this is the default)
   then `row` must contain exactly the same columns as `df` (but possibly in a
   different order).
@@ -357,9 +358,9 @@ the type of a pushed argument then a new column with a promoted element type
 allowing it is freshly allocated and stored in `df`. If `promote=false` an error
 is thrown.
 
-As a special case, if `df` has no columns and `row` is a `NamedTuple` or
-`DataFrameRow`, columns are created for all values in `row`, using their names
-and order.
+As a special case, if `df` has no columns and `row` is a `NamedTuple`,
+`DataFrameRow`, or `Tables.AbstractRow`, columns are created for all values in
+`row`, using their names and order.
 
 Please note that this function must not be used on a
 `DataFrame` that contains columns that are aliases (equal when compared with `===`).
@@ -369,7 +370,8 @@ $METADATA_FIXED
 
 """
     push!(df::DataFrame, row::Union{Tuple, AbstractArray}; promote::Bool=false)
-    push!(df::DataFrame, row::Union{DataFrameRow, NamedTuple, AbstractDict};
+    push!(df::DataFrame, row::Union{DataFrameRow, NamedTuple, AbstractDict,
+                                    Tables.AbstractRow};
           cols::Symbol=:setequal, promote::Bool=(cols in [:union, :subset]))
 
 Add one row at the end of `df` in-place, taking the values from `row`.
@@ -455,7 +457,8 @@ Base.push!(df::DataFrame, row::Any; promote::Bool=false) =
 
 """
     pushfirst!(df::DataFrame, row::Union{Tuple, AbstractArray}; promote::Bool=false)
-    pushfirst!(df::DataFrame, row::Union{DataFrameRow, NamedTuple, AbstractDict};
+    pushfirst!(df::DataFrame, row::Union{DataFrameRow, NamedTuple, AbstractDict,
+                                         Tables.AbstractRow};
                cols::Symbol=:setequal, promote::Bool=(cols in [:union, :subset]))
 
 Add one row at the beginning of `df` in-place, taking the values from `row`.
@@ -541,7 +544,8 @@ Base.pushfirst!(df::DataFrame, row::Any; promote::Bool=false) =
 
 """
     insert!(df::DataFrame, index::Integer, row::Union{Tuple, AbstractArray}; promote::Bool=false)
-    insert!(df::DataFrame, index::Integer, row::Union{DataFrameRow, NamedTuple, AbstractDict};
+    insert!(df::DataFrame, index::Integer, row::Union{DataFrameRow, NamedTuple,
+                                                      AbstractDict, Tables.AbstractRow};
             cols::Symbol=:setequal, promote::Bool=(cols in [:union, :subset]))
 
 Add one row to `df` at position `index` in-place, taking the values from `row`.
@@ -762,17 +766,20 @@ function _dfr_row_inserter!(df::DataFrame, loc::Integer, dfr::DataFrameRow,
     return _row_inserter!(df, loc, dfr, mode, cols, promote, nrows)
 end
 
-Base.push!(df::DataFrame, row::Union{AbstractDict, NamedTuple};
+Base.push!(df::DataFrame,
+           row::Union{AbstractDict, NamedTuple, Tables.AbstractRow};
            cols::Symbol=:setequal,
            promote::Bool=(cols in [:union, :subset])) =
     _row_inserter!(df, -1, row, Val{:push}(), cols, promote, -1)
 
-Base.pushfirst!(df::DataFrame, row::Union{AbstractDict, NamedTuple};
+Base.pushfirst!(df::DataFrame,
+                row::Union{AbstractDict, NamedTuple, Tables.AbstractRow};
                 cols::Symbol=:setequal,
                 promote::Bool=(cols in [:union, :subset])) =
     _row_inserter!(df, -1, row, Val{:pushfirst}(), cols, promote, -1)
 
-function Base.insert!(df::DataFrame, loc::Integer, row::Union{AbstractDict, NamedTuple};
+function Base.insert!(df::DataFrame, loc::Integer,
+                      row::Union{AbstractDict, NamedTuple, Tables.AbstractRow};
                       cols::Symbol=:setequal,
                       promote::Bool=(cols in [:union, :subset]))
     loc isa Bool && throw(ArgumentError("invalid index: $loc of type Bool"))
@@ -782,11 +789,12 @@ function Base.insert!(df::DataFrame, loc::Integer, row::Union{AbstractDict, Name
 end
 
 function _row_inserter!(df::DataFrame, loc::Integer,
-                        row::Union{AbstractDict, NamedTuple, DataFrameRow},
+                        row::Union{AbstractDict, NamedTuple, DataFrameRow,
+                                   Tables.AbstractRow},
                         mode::Union{Val{:push}, Val{:pushfirst}, Val{:insert}},
                         cols::Symbol, promote::Bool, nrows::Int)
     if nrows == -1
-        @assert row isa Union{AbstractDict, NamedTuple}
+        @assert row isa Union{AbstractDict, NamedTuple, Tables.AbstractRow}
         possible_cols = (:orderequal, :setequal, :intersect, :subset, :union)
         if !(cols in possible_cols)
             throw(ArgumentError("`cols` keyword argument must be any of :" *
@@ -800,7 +808,7 @@ function _row_inserter!(df::DataFrame, loc::Integer,
     ncols = ncol(df)
     targetrows = nrows + 1
 
-    if ncols == 0 && row isa Union{NamedTuple, DataFrameRow}
+    if ncols == 0 && row isa Union{NamedTuple, DataFrameRow, Tables.AbstractRow}
         for (n, v) in pairs(row)
             setproperty!(df, n, fill!(Tables.allocatecolumn(typeof(v), 1), v))
         end
@@ -835,11 +843,7 @@ function _row_inserter!(df::DataFrame, loc::Integer,
                 end
                 throw(AssertionError("Error adding value to column :$colname."))
             end
-            if haskey(row, colname)
-                val = row[colname]
-            else
-                val = missing
-            end
+            val = get(row, colname, missing)
             S = typeof(val)
             T = eltype(col)
             if S <: T || promote_type(S, T) <: T
@@ -977,3 +981,4 @@ function _row_inserter!(df::DataFrame, loc::Integer,
     _drop_all_nonnote_metadata!(df)
     return df
 end
+